link to codepen implementation of your jsfiddle
You are welcome to do with this (link or CodePen) as you wish.  

I have been following [The Digital Times](https://medium.com/digital-times), found Card-Kit last year and really liked it.  I offered it as a tool to my High School students who needed a way to quickly prototype cards for a make-a-game assignment.  

When I saw the update and the JSFiddle, I decided to re-implement it to CodePen as one of my #100DaysOfCode activities.

Thanks for a great tool.

janzeteachesit